Add Support for refresh rate on Windows

+++ Gems/Atom/RHI/Code/Include/Atom/RHI/Device.h
@@ -140,7 +140,7 @@ namespace AZ
             //! Get the memory requirements for allocating a buffer resource.
             virtual ResourceMemoryRequirements GetResourceMemoryRequirements(const BufferDescriptor& descriptor) = 0;
 
-            virtual float GetMainDisplayRefreshRate();
+            virtual uint32_t GetMainDisplayRefreshRate();
             
         protected:
             DeviceFeatures m_features;
